Adds superclass Contact and it is inherited by Address, Email and Phone
